+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - Update xp regeneration system, make it a separate "mana" system.
  - Fix some strange behavior with how material positions are saved when there are too many for the wand to hold.
  - When randomizing a wand, make sure it has at least enough max xp to cast its most expensive spell.
+ - Use desht's Exp manager plugin to fix some of the xp-related issues.
 
 ## 2.5.3